pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-git-bridge"
version = "0.1.0"
description = "Automate Obsidian vault Git setup for cross-device sync"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Obsidian Git Bridge", email = "dev@example.com"}
]
keywords = ["obsidian", "git", "sync", "cli", "vault"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0.0",
    "GitPython>=3.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
obsidian-git-bridge = "obsidian_git_bridge.cli:main"
ogb = "obsidian_git_bridge.cli:main"

[project.urls]
Homepage = "https://github.com/example/obsidian-git-bridge"
Repository = "https://github.com/example/obsidian-git-bridge"
Issues = "https://github.com/example/obsidian-git-bridge/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
obsidian_git_bridge = ["templates/*"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=obsidian_git_bridge --cov-report=term-missing"
